ref: move wfn info to scripts
